Chimuel 06 20 2016v2



See merge request !1

+++ database/migrations/2015_01_15_114412_create_role_user_table.php
@@ -3,23 +3,24 @@
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-class CreateRoleUserTable extends Migration
-{
+class CreateRoleUserTable extends Migration {
+
     /**
      * Run the migrations.
      *
      * @return void
      */
-    public function up()
-    {
-        Schema::create('role_user', function (Blueprint $table) {
-            $table->increments('id')->unsigned();
-            $table->integer('role_id')->unsigned()->index();
-            $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
-            $table->integer('user_id')->unsigned()->index();
-            $table->foreign('user_id')->references('user_id')->on('user')->onDelete('cascade');
-            $table->timestamps();
-        });
+    public function up() {
+        if (!Schema::hasTable('role_user')) {
+            Schema::create('role_user', function (Blueprint $table) {
+                $table->increments('id')->unsigned();
+                $table->integer('role_id')->unsigned()->index();
+                $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
+                $table->integer('user_id')->unsigned()->index();
+                $table->foreign('user_id')->references('user_id')->on('user')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -27,8 +28,8 @@ class CreateRoleUserTable extends Migration
      *
      * @return void
      */
-    public function down()
-    {
+    public function down() {
         Schema::drop('role_user');
     }
+
 }
