+++ database/migrations/2015_01_26_115523_create_permission_role_table.php
@@ -3,23 +3,24 @@
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-class CreatePermissionRoleTable extends Migration
-{
+class CreatePermissionRoleTable extends Migration {
+
     /**
      * Run the migrations.
      *
      * @return void
      */
-    public function up()
-    {
-        Schema::create('permission_role', function (Blueprint $table) {
-            $table->increments('id')->unsigned();
-            $table->integer('permission_id')->unsigned()->index();
-            $table->foreign('permission_id')->references('id')->on('permissions')->onDelete('cascade');
-            $table->integer('role_id')->unsigned()->index();
-            $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
-            $table->timestamps();
-        });
+    public function up() {
+        if (!Schema::hasTable('permission_role')) {
+            Schema::create('permission_role', function (Blueprint $table) {
+                $table->increments('id')->unsigned();
+                $table->integer('permission_id')->unsigned()->index();
+                $table->foreign('permission_id')->references('id')->on('permissions')->onDelete('cascade');
+                $table->integer('role_id')->unsigned()->index();
+                $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -27,8 +28,8 @@ class CreatePermissionRoleTable extends Migration
      *
      * @return void
      */
-    public function down()
-    {
+    public function down() {
         Schema::drop('permission_role');
     }
+
 }
