+++ database/migrations/2016_04_08_091357_alter_username_column.php
@@ -3,25 +3,21 @@
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-class AlterUsernameColumn extends Migration
-{
+class AlterUsernameColumn extends Migration {
+
     /**
      * Run the migrations.
      *
      * @return void
      */
-    public function up()
-    {
-        //
-        Schema::table('task', function (Blueprint $table) {
+    public function up() {
 
-            //if (!Schema::hasColumn('user_id'))
-            //{
+        if (!Schema::hasColumn('user_id')) {
+            Schema::table('task', function (Blueprint $table) {
                 $table->dropColumn('username');
                 $table->integer('user_id');
-            //}
-
-        });
+            });
+        }
     }
 
     /**
@@ -29,16 +25,15 @@ class AlterUsernameColumn extends Migration
      *
      * @return void
      */
-    public function down()
-    {
+    public function down() {
         //
         Schema::table('task', function (Blueprint $table) {
 
-            if (Schema::hasColumn('user_id'))
-            {
+            if (Schema::hasColumn('user_id')) {
                 $table->dropColumn('user_id');
                 $table->string('username', 50);
             }
         });
     }
+
 }
